#11 clean up packages and naming
