+++ acceptance-tests/src/test/java/de/codecentric/selenium/OpenSpaceSessionsPageTest.java
@@ -1,4 +1,4 @@
-package org.cc.thot.selenium;
+package de.codecentric.selenium;
 
 import static org.junit.Assert.assertTrue;
 
@@ -6,21 +6,21 @@ import org.junit.Test;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
-public class VenueMapPageTest extends AbstractPageTest {
+public class OpenSpaceSessionsPageTest extends AbstractPageTest {
 
-	private String pageLink = "/feedback";
+	private String pageLink = "/allSessions";
 	
 	@Test
 	public void openPageAndValidateTitle() {
 		open(pageLink);
 		assertTrue("Page title does not match", PAGE_TITLE.equalsIgnoreCase(driver.getTitle()));
 	}
-	
+
 	@Test
-	public void openVenueMapPage() {
+	public void openCurrentSessionsPage() {
 		open(pageLink);
 		WebElement title = driver.findElement(By.id("title"));
 		assertTrue(title.isDisplayed());
 	}
-
+	
 }
