+++ acceptance-tests/src/test/java/de/codecentric/selenium/StartingPageTest.java
@@ -1,4 +1,4 @@
-package org.cc.thot.selenium;
+package de.codecentric.selenium;
 
 import static org.junit.Assert.assertTrue;
 
@@ -6,21 +6,24 @@ import org.junit.Test;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
-public class CurrentSessionsPageTest extends AbstractPageTest {
+public class StartingPageTest extends AbstractPageTest {
 
-	private String pageLink = "/currentSessions";
+	private String pageLink = "/";
 	
 	@Test
 	public void openPageAndValidateTitle() {
 		open(pageLink);
 		assertTrue("Page title does not match", PAGE_TITLE.equalsIgnoreCase(driver.getTitle()));
 	}
-
+	
 	@Test
-	public void openCurrentSessionsPage() {
+	public void openStartingPage() {
 		open(pageLink);
 		WebElement title = driver.findElement(By.id("title"));
 		assertTrue(title.isDisplayed());
+
+		WebElement currentSessions = driver.findElement(By.id("current_sessions"));
+		assertTrue(currentSessions.isDisplayed());
 	}
 
 }
