+++ app/src/main/resources/application.properties
@@ -1,7 +1,6 @@
 # Spring WebApp Config
 spring.view.prefix=/WEB-INF/jsp/
 spring.view.suffix=.jsp
-# server.context-path=/thot
 
 # Database Config
 #spring.datasource.url=jdbc:mysql://192.168.59.103/ale?useUnicode=true&amp;characterEncoding=UTF-8
