+++ app/src/main/resources/static/css/conference-app.css
