+++ app/src/main/resources/static/js/conference-app.js
