+++ app/src/main/webapp/WEB-INF/jsp/footer.html
@@ -17,7 +17,7 @@
 
 <script src="<%= request.getContextPath() %>/js/jquery.js"></script>
 <script src="<%= request.getContextPath() %>/js/bootstrap.js"></script>
-<script src="<%= request.getContextPath() %>/js/thot.js"></script>
+<script src="<%= request.getContextPath() %>/js/conference-app.js"></script>
 
 </body>
 </html>
